requirements.txt
Flask==3.1.2
flask-sqlalchemy==3.1.1
requests==2.31.0
SQLAlchemy==2.0.23
APScheduler==3.10.4
python-dotenv==1.0.0
